llm_move_agent/requirements.txt
# LLM Agent 依赖

# LangChain 核心
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5

# 可选: Google Gemini 支持
# langchain-google-genai>=0.0.5

# 环境变量管理
python-dotenv>=1.0.0

# 图像处理 (摄像头)
opencv-python>=4.8.0
